+++ src/components/Icons/CloseMenuIcon/CloseMenuIcon.jsx
