+++ src/components/Icons/index.js
@@ -1,0 +1,3 @@
+export { default as BurgerMenuIcon } from "./BurgerMenuIcon/BurgerMenuIcon.jsx";
+export { default as CloseMenuIcon } from "./CloseMenuIcon/CloseMenuIcon.jsx";
+export { default as Logo } from "./Logo/Logo.jsx";
